updated proxy test
