+++ proxy/pkg/config/config.go
@@ -4,8 +4,16 @@
 package config
 
 import (
-	corev1 "k8s.io/api/core/v1"
+	"context"
+
+	"gopkg.in/yaml.v2"
+
+	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	corev1 "k8s.io/client-go/kubernetes/typed/core/v1"
+
+	"k8s.io/klog"
+	"k8s.io/kubectl/pkg/util/slice"
 )
 
 const (
@@ -46,8 +54,8 @@ func GetRBACProxyLabelMetricName() string {
 }
 
 // CreateManagedClusterLabelAllowListCM creates a managedcluster label allowlist configmap object
-func CreateManagedClusterLabelAllowListCM() *corev1.ConfigMap {
-	return &corev1.ConfigMap{
+func CreateManagedClusterLabelAllowListCM() *v1.ConfigMap {
+	return &v1.ConfigMap{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: GetManagedClusterLabelAllowListConfigMapName(),
 		},
@@ -71,4 +79,68 @@ blacklist_labels:
 - local-cluster
 - name
 `}}
+}
+
+// GetManagedClusterLabelAllowListConfigmap returns the managedcluster label allowlist configmap
+func GetManagedClusterLabelAllowListConfigmap(CoreV1Interface corev1.CoreV1Interface, namespace string) (*v1.ConfigMap,
+	error) {
+	configmap, err := CoreV1Interface.ConfigMaps(namespace).Get(
+		context.TODO(),
+		GetManagedClusterLabelAllowListConfigMapName(),
+		metav1.GetOptions{},
+	)
+	if err != nil {
+		klog.Errorf("failed to get managedcluster label allowlist configmap: %v", err)
+		return nil, err
+	}
+	return configmap, nil
+}
+
+// ModifyManagedClusterLabelAllowListConfigMapData modifies the data for managedcluster label allowlist
+func ModifyManagedClusterLabelAllowListConfigMapData(cm *v1.ConfigMap, clusterLabels map[string]string) error {
+	var clusterLabelList = &ManagedClusterLabelList{}
+
+	err := yaml.Unmarshal(
+		[]byte(cm.Data[GetManagedClusterLabelAllowListConfigMapKey()]),
+		clusterLabelList,
+	)
+	if err != nil {
+		klog.Errorf("failed to unmarshal configmap <%s> data to the clusterLabelList: %v",
+			GetManagedClusterLabelAllowListConfigMapKey(), err)
+		return err
+	}
+
+	for key := range clusterLabels {
+		if !slice.ContainsString(clusterLabelList.BlackList, key, nil) &&
+			!slice.ContainsString(clusterLabelList.LabelList, key, nil) {
+			clusterLabelList.LabelList = append(clusterLabelList.LabelList, key)
+			klog.Infof("added label <%s> to managedcluster label allowlist", key)
+		}
+	}
+
+	data, err := yaml.Marshal(clusterLabelList)
+	if err != nil {
+		klog.Errorf("failed to marshal data to clusterLabelList: %v", err)
+		return err
+	}
+	cm.Data = map[string]string{GetManagedClusterLabelAllowListConfigMapKey(): string(data)}
+	return nil
+}
+
+// UpdateManagedClusterLabelAllowListConfigMap updates the managedcluster label allowlist configmap
+func UpdateManagedClusterLabelAllowListConfigMap(
+	CoreV1Interface corev1.CoreV1Interface,
+	namespace string,
+	cm *v1.ConfigMap,
+) error {
+	_, err := CoreV1Interface.ConfigMaps(namespace).Update(
+		context.TODO(),
+		cm,
+		metav1.UpdateOptions{},
+	)
+	if err != nil {
+		klog.Errorf("failed to update managedcluster label allowlist: %v", err)
+		return err
+	}
+	return nil
 }
